+++ CHANGES.rst
@@ -10,6 +10,17 @@ trackstream
 
 - basic version test [#2]
 
+- TrackStream framework [#13]
+  The TrackStream class is the main class for creating a StreamTrack
+  given data. The TrackStream will fit the track to the data.
+
+    + In Sci-Kit Learn style with methods "fit", "predict", and "fit-predict".
+
+- StreamTrack framework [#13]
+  The StreamTrack is analogous to an interpolated CoordinateFrame
+  where the data is is a univariate function of the track arc-length.
+  The call method returns a CoordinateFrame by evaluating the interpolations.
+
 
 trackstream.tests.helper
 ^^^^^^^^^^^^^^^^^^^^^^^^